Enable usage in Ember 2.9 beta and above.
* Adds infrastructure to customize the app tree based on Ember version (at
build time).
* Fixes issues with Ember beta/canary, and removes work arounds for
  upstream issues that were fixed.
