+++ app-lt-2-9/instance-initializers/browser/head.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
 import ENV from '../../config/environment';
 
-export function initialize(owner) {
+export function initialize(instance) {
   if (ENV['ember-cli-head'] && ENV['ember-cli-head']['suppressBrowserRender']) { return true; }
 
   // clear fast booted head (if any)
@@ -9,8 +9,9 @@ export function initialize(owner) {
     .nextUntil('meta[name="ember-cli-head-end"] ~')
     .addBack()
     .remove();
-
-  const component = owner.lookup('component:head-layout');
+  const container = instance.lookup ? instance : instance.container;
+  // const renderer = container.lookup('renderer:-dom');
+  const component = container.lookup('component:head-layout');
   component.appendTo(document.head);
 }
