+++ app/instance-initializers/fastboot/head.js
@@ -1,16 +1,8 @@
-export function initialize(instance) {
-  const renderer = instance.lookup('renderer:-dom');
-  const componentFactory =
-          instance._lookupFactory('component:head-layout');
-  // explicitly set renderer & domhelper since we're in fastboot
-  const component = componentFactory.create(
-    instance.ownerInjection(),
-    {
-      renderer,
-      _domHelper: renderer._dom
-    }
-  );
-  component.appendTo(renderer._dom.document.head);
+export function initialize(owner) {
+  const document = owner.lookup('service:-document');
+  const component = owner.lookup('component:head-layout');
+
+  component.appendTo(document.head);
 }
 
 export default {
