+++ package.json
@@ -43,9 +43,11 @@
     "ember-addon"
   ],
   "dependencies": {
+    "broccoli-merge-trees": "^1.1.4",
+    "ember-cli-babel": "^5.1.5",
     "ember-cli-htmlbars": "^1.0.1",
-    "fastboot-filter-initializers": "0.0.1",
-    "ember-cli-babel": "^5.1.5"
+    "ember-cli-version-checker": "^1.1.6",
+    "fastboot-filter-initializers": "0.0.1"
   },
   "ember-addon": {
     "configPath": "tests/dummy/config"
